requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.35.0
colorama==0.4.6
decorator==4.4.2
dj3-cloudinary-storage==0.0.6
Django==4.2.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
filelock==3.12.4
future==0.18.3
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
Pillow==10.0.1
proglog==0.1.10
python-magic==0.4.27
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
